tá melhor ainda, mas temos só para aí 14.

+++ app2/src/hospital/gui/EquipmentBusyGui.java
@@ -14,7 +14,7 @@ public class EquipmentBusyGui extends JFrame {
 	
 	private Equipment myAgent;
 	
-	private int timeLeft = 10000;
+	private long timeLeft = 10000;
 	private Timer timer;
 	private JButton finishButton;
